prov/hook,verbs,rxm: Fix credit size parameter for flow ctrl
ofi_ops_flow_ctrl defines callbacks for use between rxm and
verbs.  Those callbacks include u64 parameters for credit
values.  However, the callers use size_t instead.  This
results in build warnings.

+++ prov/verbs/src/verbs_domain.c
@@ -40,7 +40,7 @@
 
 
 static void vrb_set_credit_handler(struct fid_domain *domain_fid,
-		ssize_t (*credit_handler)(struct fid_ep *ep, size_t credits))
+		ssize_t (*credit_handler)(struct fid_ep *ep, uint64_t credits))
 {
 	struct vrb_domain *domain;
